feat: implement Hyperliquid token info retrieval and refine swap example for clarity
